backend/requirements.txt
google-adk>=0.1.0
fastapi
uvicorn
pydantic
python-multipart
